% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{flatpackage_create}
\alias{flatpackage_create}
\title{Create a flatpackage}
\usage{
flatpackage_create(package, includeSuggests = TRUE)
}
\arguments{
\item{package}{Character string -  The package you want to convert to a flatpackage}

\item{includeSuggests}{Boolean - Should the packages listed in "Suggests" be 
included as well when downloading the packages for the flatpackage?}
}
\description{
The main creation function of the package. This takes a package name as input
and ultimate creates a zip file that contains everything needed to install
the package on a system.
}
\details{
System requirements aren't downloaded by default.  There will be a message
provided to tell you if any packages note any system requirements and
what those system requirements are.
}
